// workflow_fe/src/lib/api/gen/types/CreateBillingRequestDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type CreateBillingRequestDto = {
  /**
   * @type string
   */
  title: string;
  /**
   * @type string | undefined
   */
  description?: string;
  /**
   * @type string
   */
  customerName: string;
  /**
   * @type string | undefined
   */
  customerEmail?: string;
  /**
   * @type string | undefined
   */
  customerAddress?: string;
  /**
   * @minLength 0
   * @type number
   */
  amount: number;
  /**
   * @default "BDT"
   * @type string | undefined
   */
  currency?: string;
  /**
   * @type string
   */
  billingCategory: string;
  /**
   * @type string | undefined
   */
  departmentId?: string;
  /**
   * @type object | undefined
   */
  customFieldsJson?: object;
};