// workflow_fe/src/lib/api/gen/types/ResubmitExpenseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type ResubmitExpenseDto = {
  /**
   * @type string | undefined
   */
  title?: string;
  /**
   * @type string | undefined
   */
  description?: string;
  /**
   * @minLength 0
   * @type number | undefined
   */
  amount?: number;
  /**
   * @default "BDT"
   * @type string | undefined
   */
  currency?: string;
  /**
   * @type string | undefined
   */
  category?: string;
  /**
   * @type string | undefined
   */
  vendor?: string;
  /**
   * @type number | undefined
   */
  itemValue?: number;
  /**
   * @type number | undefined
   */
  price?: number;
  /**
   * @type number | undefined
   */
  quantity?: number;
  /**
   * @type string | undefined
   */
  departmentId?: string;
  /**
   * @type object | undefined
   */
  customFieldsJson?: object;
};