// src/lib/api/gen/zod/createBillingRequestDtoSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { z } from "zod/v4";

export const createBillingRequestDtoSchema = z.object({
  title: z.string(),
  description: z.optional(z.string()),
  customerName: z.string(),
  customerEmail: z.optional(z.string()),
  customerAddress: z.optional(z.string()),
  amount: z.number().min(0),
  currency: z.optional(z.string().default("BDT")),
  billingCategory: z.string(),
  departmentId: z.optional(z.string()),
  customFieldsJson: z.optional(z.object({})),
});