// workflow_fe/src/lib/api/gen/zod/resubmitBillingRequestDtoSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { z } from "zod/v4";

export const resubmitBillingRequestDtoSchema = z.object({
  title: z.optional(z.string()),
  description: z.optional(z.string()),
  customerName: z.optional(z.string()),
  customerEmail: z.optional(z.string()),
  customerAddress: z.optional(z.string()),
  amount: z.optional(z.number().min(0)),
  currency: z.optional(z.string().default("BDT")),
  billingCategory: z.optional(z.string()),
  departmentId: z.optional(z.string()),
  customFieldsJson: z.optional(z.object({})),
});